Drop support for versions of Go lower than 1.20
This commit removes the code to support Go 1.16 to 1.19, requiring now
Go 1.20. With this requirement we can remove the build tags.

It also renames the X25519 SharedKey to ECDH.
