+++ v2/go.mod
@@ -1,3 +1,3 @@
 module github.com/go-piv/piv-go/v2
 
-go 1.16
+go 1.20
